Selected which bulk capacitors will be used throughout the design, allowing for a small profile with high capacitance and fast transient response time (low ESR/ESL); also created the respective footprint
Put in a few notes regarding discrepancies between what the SIM7100 documentation suggests for cap values and what the MIC29302 datasheet says. I chose to go wivalues closer to what SIM7100 suggests but with a lower output cap value due to what photos of the mPCIe module shows. Namely, use of a few TMCUB0J107MTRF 100uF caps throughout the design, using this same cap near the regulator in place of an expensive 330uF one (as suggested in figure 9).